clean up pricing page
